requirements.txt
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
build==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Django==6.0.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
drf-spectacular==0.29.0
drf-yasg==1.21.14
Flask==3.1.2
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
pip-tools==7.5.3
pluggy==1.6.0
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==9.0.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
setuptools==82.0.0
sqlparse==0.5.5
svgwrite==1.4.3
Tree==0.2.4
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.46.3
